Split modules off `utils.defines`
The three new modules have a high internal cohesion and
a name that fixes their scope.

CMK-13768
