+++ cmk/gui/availability.py
@@ -14,7 +14,7 @@ from typing import Any, Literal, NamedTuple, Union
 
 from livestatus import LivestatusOutputFormat, OnlySites, SiteId
 
-import cmk.utils.defines as defines
+import cmk.utils.dateutils as dateutils
 import cmk.utils.paths
 import cmk.utils.store as store
 import cmk.utils.version as cmk_version
@@ -2240,23 +2240,23 @@ def _render_hour(tst: time.struct_time) -> str:
 
 
 def _render_2hours(tst: time.struct_time) -> str:
-    return defines.weekday_name(tst.tm_wday) + time.strftime(" %H:%M", tst)
+    return dateutils.weekday_name(tst.tm_wday) + time.strftime(" %H:%M", tst)
 
 
 def _render_6hours(tst: time.struct_time) -> str:
-    return defines.weekday_name(tst.tm_wday) + time.strftime(" %H:%M", tst)
+    return dateutils.weekday_name(tst.tm_wday) + time.strftime(" %H:%M", tst)
 
 
 def _render_day(tst: time.struct_time) -> str:
-    return defines.weekday_name(tst.tm_wday) + time.strftime(", %d.%m. 00:00", tst)
+    return dateutils.weekday_name(tst.tm_wday) + time.strftime(", %d.%m. 00:00", tst)
 
 
 def _render_week(tst: time.struct_time) -> str:
-    return defines.weekday_name(tst.tm_wday) + time.strftime(", %d.%m.", tst)
+    return dateutils.weekday_name(tst.tm_wday) + time.strftime(", %d.%m.", tst)
 
 
 def _render_month(tst: time.struct_time) -> str:
-    return "%s %d" % (defines.month_name(tst.tm_mon - 1), tst.tm_year)
+    return "%s %d" % (dateutils.month_name(tst.tm_mon - 1), tst.tm_year)
 
 
 def _make_struct(year: int, month: int, day: int, hour: int, *, offset: int) -> time.struct_time:
