+++ cmk/gui/bi/view.py
@@ -10,10 +10,10 @@ from typing import Any, Literal
 from livestatus import OnlySites, SiteId
 
 from cmk.utils import store
-from cmk.utils.defines import short_service_state_name
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
 from cmk.utils.servicename import ServiceName
+from cmk.utils.statename import short_service_state_name
 
 from cmk.gui.bi.bi_manager import BIManager
 from cmk.gui.bi.foldable_tree_renderer import (
