+++ cmk/gui/mkeventd/views.py
@@ -9,8 +9,8 @@ from typing import Literal, TypeGuard, TypeVar
 
 from livestatus import OnlySites, SiteId
 
-from cmk.utils.defines import short_service_state_name
 from cmk.utils.hostaddress import HostName
+from cmk.utils.statename import short_service_state_name
 from cmk.utils.user import UserId
 
 import cmk.gui.utils.escaping as escaping
