+++ cmk/gui/plugins/openapi/endpoints/time_periods/response_schemas.py
@@ -3,7 +3,7 @@
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
 
-from cmk.utils.defines import weekday_ids
+from cmk.utils.dateutils import weekday_ids
 
 from cmk.gui import fields as gui_fields
 from cmk.gui.fields.utils import BaseSchema
